Update to Ember 3.24
Some addons in the chain accidentally released a breaking change by requiring ember-modifier v3 and releasing that in a minor version.

This happened in
- @zestia/ember-auto-focus v4.1 => v4.4 went from ember-modifier v1 to v3
- ember-basic-dropdown v3.0.10 => v3.1.0 went from ember-style-modifier v0.6 to v0.7 (which requires ember-modifier v3)

Since we don't use a lock file yet, and npm 6 doesn't support overrides, we'll update to Ember 3.24 instead since it's a non-breaking change and it's supported by ember-modifier v3.